add new file in utils

+++ learnCoding/Backend/controllers/subSection.controller.js
@@ -1,7 +1,11 @@
 import mongoose from "mongoose";
 import { Section } from "../models/section.model.js";
 import { SubSection } from "../models/subSection.model.js";
-import { uploadOnCloudinary } from "../utils/cloudinary.js";
+import {
+  deleteFromCloudinary,
+  uploadOnCloudinary,
+} from "../utils/cloudinary.js";
+import { Course } from "../models/course.model.js";
 
 // create Sub section handler
 
@@ -29,9 +33,9 @@ const createSubSection = async (req, res) => {
       process.env.CLOUDINARY_VIDEO_FOLDER
     );
 
-    console.log(`uploaded video file is ${uploadedVideoFile.duration}`);
+    console.log(`uploaded video file is ${uploadedVideoFile}`);
 
-    if (!uploadedVideoFile) {
+    if (uploadedVideoFile?.result !== 'ok') {
       return res.status(400).json({
         success: false,
         message: " Error getting when upload on cloudinary",
@@ -41,12 +45,13 @@ const createSubSection = async (req, res) => {
     // save all data
     const createdSubSection = await SubSection.create({
       title,
-      timeDuration: uploadedVideoFile.duration,
+      timeDuration: uploadedVideoFile?.duration,
       description,
-      videoUrl: uploadedVideoFile.secure_url,
+      videoUrl: uploadedVideoFile?.secure_url,
     });
+    console.log(`updated section is `);
 
-    // insert sub Section in section
+    // insert sub Section id in section
     const updatedSection = await Section.findByIdAndUpdate(
       sectionId,
       {
@@ -55,7 +60,10 @@ const createSubSection = async (req, res) => {
         },
       },
       { new: true }
-    );
+    )
+      .populate("subSection")
+      .exec();
+
     console.log(`updated section is ${updatedSection}`);
 
     return res.status(200).json({
@@ -76,15 +84,17 @@ const createSubSection = async (req, res) => {
 const updateSubSection = async (req, res) => {
   try {
     // fetch data
-    const { subSectionId } = req.params.id;
-    const { title, timeDuration, description } = req.body;
-    const videoFile = req.files.video;
+    const { subSectionId, sectionId } = req.body;
+    const { title, description } = req.body;
+    const videoFile = req.files?.video;
 
-    // validate all fields
-    if (!title || !timeDuration || !description || !videoFile) {
-      return res.status(400).json({
+    console.log("video file is", videoFile);
+
+    const subSection = await SubSection.findById(subSectionId);
+    if (!subSection) {
+      return res.status(404).json({
         success: false,
-        message: " all fields are required for update sub section",
+        message: "SubSection not found",
       });
     }
 
@@ -93,9 +103,9 @@ const updateSubSection = async (req, res) => {
       videoFile,
       process.env.CLOUDINARY_VIDEO_FOLDER
     );
-    console.log(`upladed video file is ${uploadedVideoFile.secure_url}`);
+    console.log(`upladed video file is ${uploadedVideoFile?.secure_url}`);
 
-    if (!uploadedVideoFile) {
+    if (uploadedVideoFile?.result === "ok") {
       return res.status(400).json({
         success: false,
         message: " Error getting when upload on cloudinary",
@@ -106,19 +116,25 @@ const updateSubSection = async (req, res) => {
     const updatedSubSection = await SubSection.findByIdAndUpdate(
       { _id: subSectionId },
       {
-        title,
-        timeDuration,
-        description,
-        videoUrl: uploadedVideoFile.secure_url,
+        title: title && title,
+        timeDuration: uploadedVideoFile?.duration,
+        description: description && description,
+        videoUrl: uploadedVideoFile?.secure_url,
       },
       { new: true }
     );
     console.log(`something went wrong ${updatedSubSection}`);
 
+    const updatedSection = await Section.findById(sectionId)
+      .populate("subSection")
+      .exec();
+
+    console.log(`something went wrong ${updatedSection}`);
+
     return res.status(200).json({
-      success: false,
+      success: true,
       message: "Update succssfully",
-      data: updateSubSection,
+      data: updatedSection,
     });
   } catch (error) {
     console.log(`something went wrong ${error}`);
@@ -134,13 +150,14 @@ const updateSubSection = async (req, res) => {
 const deleteSubSection = async (req, res) => {
   try {
     const { subSectionId, sectionId } = req.body;
+    console.log("hello", subSectionId, sectionId);
 
     // delete id section array
     await Section.findByIdAndUpdate(
-      { _id: new mongoose.Types.ObjectId(sectionId) },
+      { _id: sectionId },
       {
         $pull: {
-          subSection: subSectionId
+          subSection: subSectionId,
         },
       },
       {
@@ -148,15 +165,36 @@ const deleteSubSection = async (req, res) => {
       }
     );
 
+    // get video url before delete subsection
+    const subSection = await SubSection.findById(subSectionId);
+    const videoUrl = subSection?.videoUrl;
+
+    console.log("hello", subSection);
+
+    // delete video from cloudinary
+    const deletedVideo = await deleteFromCloudinary(videoUrl);
+
+    if (deletedVideo?.result !== "ok") {
+      return res.status(404).json({
+        success: false,
+        message: "Could not delte video from cloudinary",
+      });
+    }
+    console.log("successfully deleted video");
+
     // delete sub section by id
     const deletedsubSection = await SubSection.findByIdAndDelete({
       _id: new mongoose.Types.ObjectId(subSectionId),
     });
 
+    const updatedSection = await Section.findById({ _id: sectionId })
+      .populate("subSection")
+      .exec();
 
     return res.status(200).json({
       success: true,
-      message: "Delete successfully seb section",
+      message: "Delete successfully sub section",
+      data: updatedSection,
     });
   } catch (error) {
     console.log(`something went wrong ${error}`);
